+++ frontend/src/components/action/bracket-styles.ts
@@ -1,0 +1,6 @@
+export const bracket_base_styles = `
+  relative inline-flex font-medium gap-[0.5px] text-brand
+  hover:after:content-[""] hover:after:absolute hover:after:bottom-[.1rem]
+  hover:after:left-[.425rem] hover:after:w-[calc(100%-.85rem)]
+  hover:after:h-px hover:after:bg-brand
+`